Added import files to facilitate cloning
To get a local copy of Shut That Down up and running, you'll need to
import the import_quotes_with_whosaid.txt so the who_said column of the
quotes table is populated (the previous import did not contain that
field).

The import to populate the quote_issues table is also included, which
will be necessary for using /browse_quotes properly.